requirements.txt
ray[serve]
unbabel-comet